+++ config/apidoc.php
@@ -15,7 +15,22 @@ return [
     /*
      * Generate a Postman collection in addition to HTML docs.
      */
-    'postman' => true,
+    'postman' => [
+        /*
+         * Specify whether the Postman collection should be generated.
+         */
+        'enabled' => true,
+
+        /*
+         * The name for the exported Postman collection. Default: config('app.name')." API"
+         */
+        'name' => null,
+
+        /*
+         * The description for the exported Postman collection.
+         */
+        'description' => null,
+    ],
 
     /*
      * The routes for which documentation should be generated.
@@ -154,4 +169,22 @@ return [
      * - size: 230 x 52
      */
     'logo' => false,
+
+    /*
+     * Configure how responses are transformed using @transformer and @transformerCollection
+     * Requires league/fractal package: composer require league/fractal
+     *
+     * If you are using a custom serializer with league/fractal,
+     * you can specify it here.
+     *
+     * Serializers included with league/fractal:
+     * - \League\Fractal\Serializer\ArraySerializer::class
+     * - \League\Fractal\Serializer\DataArraySerializer::class
+     * - \League\Fractal\Serializer\JsonApiSerializer::class
+     *
+     * Leave as null to use no serializer or return a simple JSON.
+     */
+    'fractal' => [
+        'serializer' => null,
+    ],
 ];