+++ CHANGELOG.md
@@ -13,6 +13,29 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Removed
 
+## [3.4.0] - Wednesday, 27 February 2019
+### Added
+- Support for Laravel 5.8 (https://github.com/mpociot/laravel-apidoc-generator/pull/462)
+- Ability to annotate body parameters on FormRequest (https://github.com/mpociot/laravel-apidoc-generator/pull/460)
+
+
+## [3.3.2] - Tuesday, 12 February 2019
+### Added
+- Ability to specify array and object body/query params using dot notation (https://github.com/mpociot/laravel-apidoc-generator/pull/445)
+- Ability to specify name and description of Postman collection (https://github.com/mpociot/laravel-apidoc-generator/pull/443)
+
+### Fixed
+- Postman collection and documentation base URL now uses `config('app.url')` (https://github.com/mpociot/laravel-apidoc-generator/pull/458)
+
+## [3.3.1] - Tuesday, 8 January 2019
+### Fixed
+- Fixed vendor tags (https://github.com/mpociot/laravel-apidoc-generator/pull/444)
+
+## [3.3.0] - Wednesday, 2 January 2019
+### Added
+- Ability to replace json key values in response file (https://github.com/mpociot/laravel-apidoc-generator/pull/434)
+- Support for custom transfer serializers (https://github.com/mpociot/laravel-apidoc-generator/pull/441)
+
 ## [3.2.0] - Wednesday, 12 December 2018
 ### Changed
 - API groups are now sorted "naturally" (https://github.com/mpociot/laravel-apidoc-generator/pull/428)